refactor: remove external owner compat bridges
