+++ crates/heal/src/error.rs
@@ -14,7 +14,7 @@
 
 use thiserror::Error;
 
-use super::heal::storage_compat::{DiskError, EcstoreError};
+use super::heal::{DiskError, EcstoreError};
 
 /// Custom error type for heal operations
 /// This enum defines various error variants that can occur during
