+++ crates/heal/src/heal/channel.rs
@@ -494,7 +494,7 @@ impl HealChannelProcessor {
 
 #[cfg(test)]
 mod tests {
-    use super::super::storage_compat::{DiskStore, Endpoint};
+    use super::super::{DiskStore, Endpoint};
     use super::*;
     use crate::heal::manager::HealConfig;
     use crate::heal::storage::{HealObjectInfo, HealStorageAPI};
